+++ tests/expectations/tests/macro_const.rs
@@ -7,6 +7,6 @@
 pub const foo: &'static [u8; 4usize] = b"bar\x00";
 pub const CHAR: u8 = b'b';
 pub const CHARR: u8 = b'\x00';
-pub const FLOAT: f32 = 5.09;
-pub const FLOAT_EXPR: f32 = 0.005;
+pub const FLOAT: f64 = 5.09;
+pub const FLOAT_EXPR: f64 = 0.005;
 pub const INVALID_UTF8: [u8; 5usize] = [240, 40, 140, 40, 0];
